perf(test): extract collaboration tests from startImageUpload to avoid vi.resetModules
Move the 3 collaboration-branch tests into a dedicated file where
hoisted vi.mock can replace per-test vi.resetModules + vi.doMock +
dynamic import. Tests drop from 22s to 407ms.
